"check_style_score -> 3 of 4 passed. style score: 0/1 "

+++ Module 3/3-sum/Solution.java
@@ -3,8 +3,8 @@ class solution {
 	public static void main(String[] args) {
 		Scanner input = new Scanner(System.in);
 		int n = input.nextInt();
-		int [] array;
-		array = new int[n];
+		long [] array;
+		array = new long[n];
 		for(int i= 0; i<n; i++) {
 			array[i] = input.nextInt();
 		}
@@ -12,14 +12,14 @@ class solution {
 		threesum(array, n);
 	}
 
-	public static void threesum(int[] array, int n) {
+	public static void threesum(long[] array, int n) {
 		// System.out.println(Arrays.toString(array));
 
 		int count = 0;
 		for(int i = 0; i<n-2; i++) {
 			// System.out.println("a");
 			for(int j = i+1; j<n-1; j++) {
-				int sum = 0;
+				long sum = 0;
 				// System.out.println(sum);
 				for(int k = j+1; k<n; k++) {
 					sum = array[i]+array[j]+array[k];
